Minor change as putting an if block outside the else if block.
I put an if block which updates an output value, outside the else if block, and also changed its formatting of curly brackets. Previous code was correct because the value will be changing only when we will get an alternative pair of an animal as "L" -> "Z", or else output will remain same. In order to improve its readability, I made this change.